Split translation and speech into separate tabs with enhanced speech view
Replace the combined TranslationSpeechView with a tabbed container
(TranslationStageView) that shows Translation and Speech as separate tabs.

The Speech tab includes:
- Expanded waveform display (60 bars, eager loading, seek support)
- Both base language and translated text shown per entry
- Per-item regeneration menu (three-dot dropdown)
- Word-level timecodes placeholder for future highlighting support
- Catalog type filters (text, captions, activities, answers, glossary, quizzes)
- Speech control panel with provider info and regenerate/clear actions

Also fixes missing Nepali country data in the language list.
