+++ apps/studio/src/components/pipeline/components/StepViewRouter.tsx
@@ -9,7 +9,7 @@ import {
   CaptionsView,
   GlossaryView,
   TocView,
-  TranslationSpeechView,
+  TranslationStageView,
   SpeechView,
   PreviewView,
   ValidationView,
@@ -54,7 +54,7 @@ const VIEW_MAP: Record<string, ViewEntry> = {
   captions: { component: CaptionsView },
   glossary: { component: GlossaryView },
   toc: { component: TocView },
-  translation: { component: TranslationSpeechView, fullHeight: true },
+  translation: { component: TranslationStageView, fullHeight: true },
   speech: { component: SpeechView, fullHeight: true },
   validation: { component: ValidationView, fullHeight: true },
   preview: { component: PreviewView, fullHeight: true },
