+++ apps/studio/src/components/pipeline/stages/index.ts
@@ -7,7 +7,7 @@ export { GlossaryView } from "./glossary/GlossaryView"
 export { TocView } from "./toc/TocView"
 export { TranslationsView } from "./translations/TranslationsView"
 export { SpeechView } from "./speech/SpeechView"
-export { TranslationSpeechView } from "./translations/TranslationSpeechView"
+export { TranslationStageView } from "./translations/TranslationStageView"
 export { PreviewView } from "./PreviewView"
 export { ValidationView } from "./ValidationView"
 export { ExportView } from "./ExportView"
